backend/requirements.txt
asgiref==3.7.2
bitarray==2.9.2
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.1.0
django-rest-durin==1.1.0
djangorestframework==3.14.0
drf-api-logger==1.1.14
humanize==4.7.0
idna==3.4
pdfreader==0.1.12
pillow==10.3.0
psycopg2-binary==2.9.6
pycryptodome==3.20.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
urllib3==2.0.3
webencodings==0.5.1
